- made the corpse queue a simple array in FLevelLocals.
I have to wonder why it had to use such a complicated implementation that provided no advantages whatsoever.
The new code is just 1/5th of the old one's size and much closer to Hexen's original implementation which also was a simple array but with no means to resize the queue.

# Conflicts:
#	src/g_level.cpp
#	src/g_statusbar/shared_sbar.cpp
#	src/p_saveg.cpp
#	src/p_setup.cpp